requirements.txt
discord.py
colorlog
aiosqlite
riotwatcher